Move file to base api and update auth to post only
