--- src/common/lib/api/apiCollections/index.js
@@ -1,8 +1,0 @@
-export * from "./categories"
-export * from "./debio-balance"
-export * from "./emr"
-export * from "./labs"
-export * from "./location"
-export * from "./orders"
-export * from "./rating"
-export * from "./service-request"
