+++ src/common/lib/api/gcs.js
@@ -12,10 +12,17 @@ export async function getSignedUrl(fileName, action = "read") {
 }
 
 export async function createSyncEvent({ orderId, serviceCategoryId, fileName }) {
-  const { data } = await apiClientRequest.post("/bounty/create-sync-event", {
+  const payload = {
     order_id: orderId,
     service_category_id: serviceCategoryId,
     filename: fileName
+  }
+
+  const { data } = await apiClientRequest.post("/bounty/create-sync-event", payload, {
+    auth: {
+      username: process.env.VUE_APP_USERNAME,
+      password: process.env.VUE_APP_PASSWORD
+    }
   })
 
   return data
