+++ src/common/lib/api/index.js
@@ -1,13 +1,21 @@
 import axios from "axios"
-export * from "./apiCollections"
 
+// EXPORT API COLLECTIONS HERE
+export * from "./categories"
+export * from "./debio-balance"
+export * from "./emr"
+export * from "./gcs"
+export * from "./labs"
+export * from "./location"
+export * from "./orders"
+export * from "./rating"
+export * from "./service-request"
+
+// AXIOS INSTANCE EXPORT BY DEFAULT
+// PLEASE DISCUSS BEFORE YOU WANT TO EDIT THIS SCRIPT
 const apiClientRequest = axios.create({
   baseURL: process.env.VUE_APP_BACKEND_API,
-  headers: { "Content-Type": "application/json" },
-  auth: {
-    username: process.env.VUE_APP_USERNAME,
-    password: process.env.VUE_APP_PASSWORD
-  }
+  headers: { "Content-Type": "application/json" }
 })
 
 export default apiClientRequest
