+++ src/common/lib/api/rating.js
@@ -1,13 +1,20 @@
 import apiClientRequest from "@/common/lib/api";
 
 export const submitRatingOrder = async (labId, serviceId, orderId, ratingBy, rating, review) => {
-  const { data } = await apiClientRequest.post("/rating", {
+  const payload = {
     lab_id: labId,
     service_id: serviceId,
     order_id: orderId,
     rating_by: ratingBy,
     rating: rating,
     review: review
+  }
+
+  const { data } = await apiClientRequest.post("/rating", payload, {
+    auth: {
+      username: process.env.VUE_APP_USERNAME,
+      password: process.env.VUE_APP_PASSWORD
+    }
   });
 
   return data;
