+++ src/views/Dashboard/Customer/Home/MyTest/index.vue
@@ -132,7 +132,7 @@ import CryptoJS from "crypto-js"
 import localStorage from "@/common/lib/local-storage"
 import { u8aToHex } from "@polkadot/util"
 import { syncDecryptedFromIPFS } from "@/common/lib/ipfs"
-import { createSyncEvent } from "@/common/lib/ipfs/gcs"
+import { createSyncEvent } from "@/common/lib/api"
 import { getCategories } from "@/common/lib/api"
 import { getOrdersData } from "@/common/lib/polkadot-provider/query/orders"
 import { queryLabsById } from "@/common/lib/polkadot-provider/query/labs"
