+++ src/views/LandingPage/GenerateAccount/SetPassword.vue
@@ -131,7 +131,12 @@ export default {
     },
     
     async onVerifyRecaptcha(response) {
-      const result = await apiClientRequest.post("/recaptcha", { response })
+      const result = await apiClientRequest.post("/recaptcha", { response }, {
+        auth: {
+          username: process.env.VUE_APP_USERNAME,
+          password: process.env.VUE_APP_PASSWORD
+        }
+      })
 
       if (result.data.success) this.recaptchaVerified = true
     }
