+++ src/views/LandingPage/SignIn/ChangePassword.vue
@@ -137,7 +137,12 @@ export default {
     },
     
     async onVerifyRecaptcha(response) {
-      const result = await apiClientRequest.post("/recaptcha", { response })
+      const result = await apiClientRequest.post("/recaptcha", { response }, {
+        auth: {
+          username: process.env.VUE_APP_USERNAME,
+          password: process.env.VUE_APP_PASSWORD
+        }
+      })
 
       if (result.data.success) this.recaptchaVerified = true
     }
